# openapi.yaml
openapi: 3.0.3
info:
  title: Walrus Messenger Backend API
  version: 0.1.0
  description: |
    Implemented HTTP API only (current server state).
    Request bodies larger than 64 KiB are rejected with HTTP 413.
servers:
  - url: http://127.0.0.1:3000
    description: Local default (pass `--address 0.0.0.0:3000` on startup)
tags:
  - name: auth
  - name: messaging
paths:
  /health:
    get:
      tags: [auth]
      summary: Public liveness check
      operationId: health
      description: Lightweight health check endpoint without authentication.
      security: []
      responses:
        '200':
          description: Service is up

  /auth/login:
    post:
      tags: [auth]
      summary: Login and create session
      operationId: login
      description: >
        Authenticates user alias/password and returns access + refresh tokens.
        Request field `session_id` is accepted by the server model but currently unused.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthPayload'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenExchangePayload'
        '401':
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: bad auth or refresh credentials
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/whoami:
    get:
      tags: [auth]
      summary: Resolve current user by access token
      operationId: whoAmI
      description: Validates bearer access token and returns authenticated user id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhoAmIResponse'
        '400':
          description: Missing or malformed bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Missing or bad token in request
        '401':
          description: Token expired or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh session tokens
      operationId: refreshSession
      description: >
        Rotates both access and refresh tokens using current refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshPayload'
      responses:
        '200':
          description: Tokens rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenExchangePayload'
        '401':
          description: Bad or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Concurrent refresh interrupted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags: [auth]
      summary: Logout current session
      operationId: logoutSession
      description: Invalidates the current authenticated session.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Session invalidated
        '400':
          description: Missing or malformed bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token expired or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/change-password:
    post:
      tags: [auth]
      summary: Change current user password
      operationId: changePassword
      description: Validates current password and updates it to a new value.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          description: Missing or malformed bearer token, or invalid new password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token expired or current password mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/invite:
    post:
      tags: [auth]
      summary: Invite user and bootstrap direct chats
      operationId: inviteUser
      description: >
        Admin-only endpoint. Creates a regular user from alias/password and immediately creates
        private chats between that user and every existing user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteUserResponse'
        '400':
          description: Invalid payload, insufficient permissions, or user alias already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token expired or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chats:
    get:
      tags: [messaging]
      summary: List chats for current user
      operationId: listChats
      description: >
        Returns paginated chats visible to current user.
        `display_name` is normalized for UI list usage; for private chats it resolves to peer display name.
        Uses page mode parameters: `limit` and `page`.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 100
        - in: query
          name: page
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 1
      responses:
        '200':
          description: Chats page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListChatsResponse'
        '400':
          description: Invalid query params or malformed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token expired or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chats/{chat_id}/messages:
    get:
      tags: [messaging]
      summary: List messages in a chat
      operationId: listMessages
      description: >
        Returns messages for a chat if current user is a member.
        With `offset`, response contains messages with IDs greater than it (incremental mode).
        Without `offset`, regular page mode (`limit` + `page`) is used.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 100
        - in: query
          name: page
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 1
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Messages page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMessagesResponse'
        '400':
          description: Invalid query params or malformed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token expired or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found or user has no access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [messaging]
      summary: Send message to a chat
      operationId: sendMessage
      description: Creates a new text message in the chat if current user is a member.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          description: Invalid payload or malformed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token expired or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found or user has no access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaqueBase64SessionToken
  schemas:
    AuthPayload:
      type: object
      additionalProperties: false
      required: [alias, password]
      properties:
        alias:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        session_id:
          type: string
          nullable: true
          description: Currently accepted but not used by backend logic.

    RefreshPayload:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
          description: Opaque refresh token returned by /auth/login or /auth/refresh.

    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 1
        new_password:
          type: string
          minLength: 8
          maxLength: 80

    InviteUserRequest:
      type: object
      additionalProperties: false
      required: [alias, password]
      properties:
        alias:
          type: string
          minLength: 1
          maxLength: 30
        password:
          type: string
          minLength: 8
          maxLength: 80

    InviteUserResponse:
      type: object
      additionalProperties: false
      required: [user_id]
      properties:
        user_id:
          type: integer
          format: int32

    WhoAmIResponse:
      type: object
      additionalProperties: false
      required: [user_id]
      properties:
        user_id:
          type: integer
          format: int32

    TokenExchangePayload:
      type: object
      additionalProperties: false
      required:
        - refresh_token
        - refresh_token_expires_at
        - access_token
        - access_token_expires_at
      properties:
        refresh_token:
          type: string
          description: Opaque token string.
        refresh_token_expires_at:
          type: string
          format: date-time
        access_token:
          type: string
          description: Opaque token string.
        access_token_expires_at:
          type: string
          format: date-time

    ChatKind:
      type: string
      enum: [with_self, private, group, channel]

    ChatResponse:
      type: object
      additionalProperties: false
      required: [id, display_name, kind]
      properties:
        id:
          type: integer
          format: int64
        display_name:
          type: string
          nullable: true
        kind:
          $ref: '#/components/schemas/ChatKind'

    ListChatsResponse:
      type: object
      additionalProperties: false
      required: [chats]
      properties:
        chats:
          type: array
          items:
            $ref: '#/components/schemas/ChatResponse'

    MessageResponse:
      type: object
      additionalProperties: false
      required: [id, text, created_at, edited_at, user_id, user_display_name]
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        edited_at:
          type: string
          format: date-time
          nullable: true
        user_id:
          type: integer
          format: int32
          nullable: true
        user_display_name:
          type: string
          nullable: true

    ListMessagesResponse:
      type: object
      additionalProperties: false
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageResponse'

    SendMessageRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4096

    SendMessageResponse:
      type: object
      additionalProperties: false
      required: [message_id]
      properties:
        message_id:
          type: integer
          format: int64

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string